pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "futbot"
version = "0.1.0"
description = "Automated Turkish football news bot for X/Twitter"
authors = [{name = "Football Bot", email = "bot@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.10",
    "trafilatura>=1.6.4",
    "beautifulsoup4>=4.12.2",
    "httpx[http2]>=0.25.0",
    "tldextract>=5.0.1",
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.0",
    "langdetect>=1.0.9",
    "rapidfuzz>=3.5.2",
    "readability-lxml>=0.8.1",
    "sumy>=0.11.0",
    "nltk>=3.8.1",
    "Pillow>=10.1.0",
    "requests>=2.31.0",
    "tweepy>=4.14.0",
    "pytz>=2023.3",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "schedule>=1.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6"
]

[project.scripts]
futbot = "app.main:app"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
